split shared + taxCode utils
