+++ src/components/BankTransactions/BankTransactionsHeader.tsx
@@ -18,10 +18,10 @@ import { useBankTransactionsFiltersContext } from '@contexts/BankTransactionsFil
 import { useBankTransactionsIsCategorizationEnabledContext } from '@contexts/BankTransactionsIsCategorizationEnabledContext/BankTransactionsIsCategorizationEnabledContext'
 import { HStack, VStack } from '@ui/Stack/Stack'
 import { Toggle } from '@ui/Toggle/Toggle'
+import { BankTransactionsDateFilterMode } from '@components/BankTransactions/bankTransactions'
 import { BankTransactionsBulkActions } from '@components/BankTransactions/BankTransactionsBulkActions/BankTransactionsBulkActions'
 import { BankTransactionsHeaderMenu, BankTransactionsHeaderMenuActions } from '@components/BankTransactions/BankTransactionsHeaderMenu'
 import { BankTransactionsTableContent } from '@components/BankTransactions/constants'
-import { BankTransactionsDateFilterMode } from '@components/BankTransactions/bankTransactions'
 import { BankTransactionsActions } from '@components/BankTransactionsActions/BankTransactionsActions'
 import { BulkActionsModule } from '@components/BulkActionsModule/BulkActionsModule'
 import { ButtonVariant } from '@components/Button/Button'
