+++ src/components/BankTransactionsMobileList/BankTransactionsMobileListItem.tsx
@@ -14,10 +14,10 @@ import FileIcon from '@icons/File'
 import { AnimatedPresenceElement } from '@ui/AnimatedPresenceElement/AnimatedPresenceElement'
 import { HStack, VStack } from '@ui/Stack/Stack'
 import { Span } from '@ui/Typography/Text'
-import { BankTransactionsAmountDate } from '@components/BankTransactions/BankTransactionsAmountDate'
-import { BankTransactionsListItemCategory } from '@components/BankTransactions/BankTransactionsListItemCategory/BankTransactionsListItemCategory'
 import { hasReceipts, isCredit } from '@components/BankTransactions/bankTransactions'
 import { isCategorized } from '@components/BankTransactions/bankTransactions'
+import { BankTransactionsAmountDate } from '@components/BankTransactions/BankTransactionsAmountDate'
+import { BankTransactionsListItemCategory } from '@components/BankTransactions/BankTransactionsListItemCategory/BankTransactionsListItemCategory'
 import { BankTransactionsProcessingInfo } from '@components/BankTransactionsList/BankTransactionsProcessingInfo'
 import { BankTransactionsMobileListItemCheckbox } from '@components/BankTransactionsMobileList/BankTransactionsMobileListItemCheckbox'
 import { BankTransactionsMobileListItemExpandedRow } from '@components/BankTransactionsMobileList/BankTransactionsMobileListItemExpandedRow'
