+++ src/hooks/utils/bankTransactions/shared.ts
@@ -1,12 +1,12 @@
 import { isWithinInterval, parseISO } from 'date-fns'
 
-import { type BankTransaction, type BankTransactionTaxOption, DisplayState, type Split, type SuggestedMatch } from '@internal-types/bankTransactions'
+import { type BankTransaction, DisplayState, type Split, type SuggestedMatch } from '@internal-types/bankTransactions'
 import { type DateRange } from '@internal-types/general'
 import { Direction } from '@internal-types/general'
 import type { TagFilterInput } from '@internal-types/tags'
 import type { CategoryUpdate } from '@schemas/bankTransactions/categoryUpdate'
-import type { Classification } from '@schemas/categorization'
 import { makeTagKeyValueFromTag } from '@schemas/tag'
+import { getCategoryPayloadTaxCode } from '@hooks/utils/bankTransactions/taxCode'
 import { CategorizedCategories, ReviewCategories } from '@components/BankTransactions/constants'
 
 export const filterVisibility = (
@@ -77,60 +77,6 @@ export const countTransactionsToReview = ({
 
 export const hasReceipts = (bankTransaction?: BankTransaction) =>
   bankTransaction?.document_ids && bankTransaction.document_ids.length > 0
-
-export type BankTransactionTaxCodeOption = {
-  label: string
-  value: string
-}
-
-export const getBankTransactionTaxOptions = (bankTransaction?: BankTransaction): BankTransactionTaxOption[] => {
-  if (!bankTransaction?.tax_options) {
-    return []
-  }
-
-  return Object.values(bankTransaction.tax_options).flat()
-}
-
-export const getBankTransactionTaxCodeOptions = (bankTransaction?: BankTransaction): BankTransactionTaxCodeOption[] => {
-  return getBankTransactionTaxOptions(bankTransaction).map(taxOption => ({
-    label: taxOption.display_name,
-    value: taxOption.code,
-  }))
-}
-
-export const getBankTransactionTaxCodeOption = (
-  bankTransaction: BankTransaction | undefined,
-  taxCode: string | null,
-): BankTransactionTaxCodeOption | null => {
-  if (!taxCode) {
-    return null
-  }
-
-  const taxOption = getBankTransactionTaxOptions(bankTransaction).find(option => option.code === taxCode)
-
-  return {
-    label: taxOption?.display_name ?? taxCode,
-    value: taxCode,
-  }
-}
-
-export const hasBankTransactionTaxCode = (
-  bankTransaction: BankTransaction | undefined,
-  taxCode: string | null,
-) => {
-  return getBankTransactionTaxOptions(bankTransaction).some(taxOption => taxOption.code === taxCode)
-}
-
-export const getCategoryPayloadTaxCode = (
-  classification: Classification | null | undefined,
-  taxCode: string | null | undefined,
-) => {
-  if (classification?.type === 'Exclusion') {
-    return null
-  }
-
-  return taxCode ?? null
-}
 
 export const isTransferMatch = (bankTransaction?: BankTransaction) => {
   return bankTransaction?.match?.details.type === 'Transfer_Match'
